+++ app/scripts/lib/lforms-fhir-sdc.js
@@ -1,15 +1,15 @@
 /**
- * A package to handle FHIR SDC (STU2 Ballot, Version 1.6.0) Questionnaire and QuestionnaireResponse for LForms
+ * A package to handle FHIR SDC (STU2) Questionnaire and QuestionnaireResponse for LForms
  * STU2 Ballot:
- * http://hl7.org/fhir/us/sdc/2016Sep/sdc-questionnaire.html
- * http://hl7.org/fhir/us/sdc/2016Sep/sdc-response.html
+ * http://hl7.org/fhir/us/sdc/sdc-questionnaire.html
+ * http://hl7.org/fhir/us/sdc/sdc-questionnaireresponse.html
  *
  * It provides the following functions:
- * convert2Questionnaire()
+ * convertLFormsToQuestionnaire()
  * -- Convert existing LOINC panels/forms data in LForms format into FHIR SDC Questionnaire data
- * convert2QuestionnaireResponse()
+ * convertLFormsToQuestionnaireResponse()
  * -- Generate FHIR SDC QuestionnaireResponse data from captured data in LForms
- * mergeQuestionnaireResponseToForm()
+ * mergeQuestionnaireResponseToLForms()
  * -- Merge FHIR SDC QuestionnaireResponse data into corresponding LForms data
  */
 if (typeof LForms === 'undefined')
@@ -25,7 +25,7 @@ jQuery.extend(LForms.FHIR_SDC, {
    * @param lfData a LForms form object
    * @returns {{}}
    */
-  convert2Questionnaire: function(lfData) {
+  convertLFormsToQuestionnaire: function(lfData) {
     var target = {};
 
     if (lfData) {
@@ -471,7 +471,7 @@ jQuery.extend(LForms.FHIR_SDC, {
    * @param lfData a LForms form object
    * @returns {{}}
    */
-  convert2QuestionnaireResponse: function(lfData) {
+  convertLFormsToQuestionnaireResponse: function(lfData) {
     var target = {};
     if (lfData) {
       var source = lfData.getFormData(true,true,true,true);
@@ -1111,10 +1111,13 @@ jQuery.extend(LForms.FHIR_SDC, {
    * @param qr a QuestionnaireResponse instance
    * @returns {{}} an updated LForms form object
    */
-  mergeQuestionnaireResponseToForm : function(formData, qr) {
+  mergeQuestionnaireResponseToLForms : function(formData, qr) {
+
+    // get the default settings in case they are missing in the form data
+    var newFormData = (new LForms.LFormsData(formData)).getFormData();
     var qrInfo = this._getQRStructure(qr);
-    this._processQRItemAndLFormsItem(qrInfo, formData);
-    return formData;
+    this._processQRItemAndLFormsItem(qrInfo, newFormData);
+    return newFormData;
   },